Move `handleModalLoanReservation` out of `MaterialButtonsOnlineInternal`
Added a custom hook called `useOnlineInternalHandleLoanReservation` to improve the readability of `MaterialButtonsOnlineInternal`.
