+++ src/components/material/material-buttons/online/MaterialButtonsOnlineInternal.tsx
@@ -1,5 +1,4 @@
 import React, { FC, useState } from "react";
-import { useQueryClient } from "react-query";
 import { Manifestation } from "../../../../core/utils/types/entities";
 import MaterialSecondaryLink from "../generic/MaterialSecondaryLink";
 import MaterialSecondaryButton from "../generic/MaterialSecondaryButton";
@@ -10,24 +9,13 @@ import { ButtonSize } from "../../../../core/utils/types/button";
 import useReaderPlayer from "../../../../core/utils/useReaderPlayer";
 import LinkButton from "../../../Buttons/LinkButton";
 import { Button } from "../../../Buttons/Button";
-import {
-  getGetV1UserLoansQueryKey,
-  getGetV1UserReservationsQueryKey,
-  usePostV1UserLoansIdentifier,
-  usePostV1UserReservationsIdentifier
-} from "../../../../core/publizon/publizon";
-import {
-  getAllFaustIds,
-  getManifestationType
-} from "../../../../core/utils/helpers/general";
-import { useUrls } from "../../../../core/utils/url";
-import { onlineInternalModalId } from "../../../../apps/material/helper";
-import { usePatronData } from "../../../../core/utils/helpers/usePatronData";
+import { getManifestationType } from "../../../../core/utils/helpers/general";
 import { OnlineInternalRequestStatus } from "../../../../core/utils/types/request";
 import DeleteReservationModal, {
   deleteReservationModalId
 } from "../../../../apps/reservation-list/modal/delete-reservation/delete-reservation-modal";
 import { ReservationType } from "../../../../core/utils/types/reservation-type";
+import useOnlineInternalHandleLoanReservation from "../../../../core/utils/useOnlineInternalHandleLoanReservation";
 
 type MaterialButtonsOnlineInternalType = {
   size?: ButtonSize;
@@ -44,13 +32,8 @@ const MaterialButtonsOnlineInternal: FC<MaterialButtonsOnlineInternalType> = ({
   openModal,
   setReservationStatus
 }) => {
-  const queryClient = useQueryClient();
   const t = useText();
-  const u = useUrls();
-  const authUrl = u("authUrl");
-  const { open, openGuarded } = useModalButtonHandler();
-  const { mutate: mutateLoan } = usePostV1UserLoansIdentifier();
-  const { mutate: mutateReservation } = usePostV1UserReservationsIdentifier();
+  const { open } = useModalButtonHandler();
   const {
     type,
     orderId,
@@ -61,72 +44,13 @@ const MaterialButtonsOnlineInternal: FC<MaterialButtonsOnlineInternalType> = ({
     canBeReserved,
     reservation
   } = useReaderPlayer(manifestations);
-  const { data: userData } = usePatronData();
+  const handleModalLoanReservation = useOnlineInternalHandleLoanReservation({
+    manifestations,
+    openModal,
+    setReservationStatus
+  });
   const [reservationToDelete, setReservationToDelete] =
     useState<ReservationType | null>(null);
-
-  const handleModalLoanReservation = () => {
-    if (openModal) {
-      openGuarded({
-        authUrl,
-        modalId: onlineInternalModalId(getAllFaustIds(manifestations))
-      });
-      return;
-    }
-
-    if (canBeLoaned && identifier) {
-      mutateLoan(
-        { identifier },
-        {
-          onSuccess: () => {
-            // Ensure that the button is updated after a successful loan
-            queryClient.invalidateQueries(getGetV1UserLoansQueryKey());
-            if (setReservationStatus) {
-              setReservationStatus("loaned");
-            }
-          },
-          onError: () => {
-            if (setReservationStatus) {
-              setReservationStatus("error");
-            }
-          }
-        }
-      );
-      return;
-    }
-
-    if (canBeReserved && identifier && userData?.patron) {
-      mutateReservation(
-        {
-          identifier,
-          data: {
-            email: userData.patron.emailAddress,
-            // Only add phone number if it exists
-            // Still waiting for the API to support optional phoneNumber
-            ...(userData.patron.phoneNumber && {
-              phoneNumber: userData.patron.phoneNumber.match(/^\+\d{2}/)
-                ? userData.patron.phoneNumber // Keep the number unchanged if it already starts with +XX
-                : `+45${userData.patron.phoneNumber}` // Prepend +45 if no country code is present
-            })
-          }
-        },
-        {
-          onSuccess: () => {
-            // Ensure that the button is updated after a successful reservation
-            queryClient.invalidateQueries(getGetV1UserReservationsQueryKey());
-            if (setReservationStatus) {
-              setReservationStatus("reserved");
-            }
-          },
-          onError: () => {
-            if (setReservationStatus) {
-              setReservationStatus("error");
-            }
-          }
-        }
-      );
-    }
-  };
 
   const manifestationType = getManifestationType(manifestations);
   const reseveLabel = openModal
